com essa branch se tornou possivel subir uma imagem como capa do banner, salvar ela, recuperar ela, atualizar a URL dela na hora de receber o post, a rota sabe se tem uma imagem de capa ja salva,  e se tiver dela a anterior e sava a nova, assim como é possível apenas deletar e salvar o post sem imagem de capa.

+++ back-end/src/Modules/Posts/useCases/uploadFeatureImage/uploadFeatureImageUseCase.ts
@@ -27,12 +27,15 @@ class UploadFeatureImageUseCase {
     async execute({image_name, article_id, folder}: IRequest): Promise<string> {
         const articleExists = await this.articleRepo.findById(article_id)
 
+        if(!articleExists) {
+            throw new Error("This article does not exists").message
+        }
+
         if(articleExists.feature_image) {   
             const image = await this.articleImageRepo.findById(articleExists.id)
             
             await this.storageProvider.delete(image_name, image.folder)
             await this.articleImageRepo.delete(image.image_name)
-            console.log("deleted")
         }
 
         let now = new Date();
@@ -44,14 +47,8 @@ class UploadFeatureImageUseCase {
         }
 
         const folderDest = `content/images/${year}/${month}/${folder}/`
-        
-
-        if(!articleExists) {
-            throw new Error("This article does not exists").message
-        }
 
         const feature_image = await this.storageProvider.save(image_name, "images")
-        console.log(feature_image)
 
         await this.articleImageRepo.create(
             image_name,
